+++ buildSrc/src/main/java/Config.kt
@@ -33,11 +33,11 @@ object Config {
     object Libs {
         val kotlinStd = "org.jetbrains.kotlin:kotlin-stdlib"
 
-        val sentryJavaVersion = "8.34.1"
+        val sentryJavaVersion = "8.41.0"
         val sentryAndroid = "io.sentry:sentry-android:$sentryJavaVersion"
         val sentryJava = "io.sentry:sentry:$sentryJavaVersion"
 
-        val sentryCocoaVersion = "8.57.3"
+        val sentryCocoaVersion = "8.58.2"
         val sentryCocoa = "Sentry"
 
         object Samples {