+++ sentry-kotlin-multiplatform/src/commonJvmMain/kotlin/io/sentry/kotlin/multiplatform/extensions/SentryOptionsExtensions.jvm.kt
@@ -67,4 +67,7 @@ internal fun JvmSentryOptions.applyJvmBaseOptions(kmpOptions: SentryOptions) {
             }
         }
     }
+    kmpOptions.proguardUuid?.let { uuid ->
+        jvmOptions.proguardUuid = uuid
+    }
 }